Save the NavBackStackEntry state
When using the saveState flag on the pop operation,
all of the state associated with the popped destinations
should be saved. That includes both the saved
instance state (calling saveState() once the destination
is stopped) as well as non-config state in the form of
any ViewModels associated with the destination.

This involves associating that saved state with
the ID of the destination you've popped up to
as well as ensuring that ViewModels are not cleared
as part of the pop (but are still cleared when the
hosting ViewModelStore is cleared).

Relnote: N/A
Test: new NavBackStackEntryTest tests pass
BUG: 80029773
Change-Id: Ia5a35ec712cf5a58c3c310999bacb3c87e8cc522