Removing support for moving focus to the parent using back key
DemoTest uses the back key to exit a demo. aosp/1666191 added a feature where
the back button could be used to move focus to a parent. So pressing back would
clear focus (assign focus to the root) if the focused item did not have a parent.
This broke the DemoTests because we consumed the back event.

This CL makes the focus system ignore back events for now.

Bug: 185211677
Bug: 183746743
Test: androidx.compose.integration.demos.test.DemoTest
Change-Id: Iff99cd445b94a667eb2ad5e4090902c58842148f