WIP add support for Firebase Test Lab
This CL introduces a basic setup in AndroidX to run instrumentation
tests on Firebase Test Lab.

This is limited to only application tests for now because FTL requires
two APKs. We can probably work around it if we want to start using FTL
extensively.

This PR only enables tests in Room for now but tasks are created for all
projects and can also be run locally if gcloud console is initialized
(in playground setup).

Bug: n/a
Test: not sure how to test this so for now, none :).