Expand FTL to fetch results into dist
This CL adds support for specifying a Bucket to keep results, which
defaults to a generic bucket I've created for FTL tests.

By default, the bucket path is inferred from github environment
variables when available, otherwise, a random UUID is chosen for uniqueness.

FTL outputs a bunch of things into this bucket, including a video of the
test running. All of it is pulled back into the out folder after the
test runs. I've also configured a finalizer task to copy that output
into the DIST folder.

Since we only rely on APK contents, I marked the task cacheable.

Bug: n/a
Test: production :(. verified build cache locally but required
--build-cache