Adds configuration to define processor target language
Adds configuration that defines if a processing environment should
expect to work as if on a regular Java setup or if it's Kotlin only.
This defines internal behavior as to how some things will be modelled,
for example if we can expect top level methods to be modelled as such,
or if we should synthesize a `Kt` type to hold it instead.

For JavaAP, the `XProcessingStep.asAutoCommonProcessor()` method now
expects the language to be specified. If not specified, it will default
to expect Java language. KSP only expects Kotlin language.