Add extra supported surface combinations quirk
The guaranteed supported configurations tables only list the minimal requirements of the stream combinations for each hardware level of devices. In fact, some devices might be able to support more stream combinations. This quirk allows CameraX to add the additional supported surface combinations so that it will benefit when binding more use cases or using YUV stream for ImageCapture.

Bug: 183478297
Test: ./gradlew camera:camera-camera2:build

Change-Id: I027f3cb1e0176e379add37e40ee7c8d8e9766930