Model top level functions and properties
This CL adds support for top level functions/properties in KSP.

KSP does not generate synthetic class for files that include top
level methods or properites.
see: https://github.com/google/ksp/issues/396

For XProcessing, i decided the follow through and not create a fake
XTypeElement for them. Instead, this CL introduces a new XElement
type called XMemberContainer. This interface defines two properties:
className:ClassName and type:XType.
ClassName is mandatory whereas XType is optional (to denote that it
does not have an accessible type in KSP but ClassName can be accessed
for code generation).
This was enough to make the rest of the codebase happy.

For Room, I also removed `XMethod/FieldElement.requireEnclosingTypeElement`
as we didn't really need it once we have className.

XMemberContainer might be extended to provide access to more declarations
but we don't yet have a use case for that.

By not modeling fake XTypeElements, we avoid creating a fake XType for it,
which prevents lots of pitfalls since the type does not really exist in
kotlin world.

Bug: 160322705
Fixes: 184526463
Test: XRoundEnvTest, TopLevelMembersTest
Change-Id: Iea60d2587c9373a96a56f7721d5ea2782b66ba2e