Enable non-interactive Compose tests on Robolectric
On Robolectric, idling resources are not awaited, draw passes don't
happen, and the whole test runs on the main thread. Compose tests, on
the other hand, rely on idling resources, draw passes, and not running
on the main thread.

= Main thread issue =
The main thread issue is that we must run Espresso.onIdle() from another
thread than the main thread (or it will deadlock), while on Robolectric
everything is run from the main thread, so that includes the call to
Espresso.onIdle(). Our are-you-on-the-main-thread checks therefor need
to be inverted on Robolectric.
To fix this, we now have an IdlingStrategy interface that implements the
run until idle functionality, and does the right check or thread jump.
External test frameworks can change this behavior by installing a
factory for the IdlingStrategy before the tests commence.

= Idling resources issue =
The idling resources issue is that we run our synchronization from an
IdlingResource that we install into Espresso. Espresso doesn't run
IdlingResources, so our synchronization is never run.
To fix this issue, the RobolectricIdlingStrategy specifically queries
ComposeIdlingResource after Espresso.onIdle(). We keep alternating these
two methods until ComposeIdlingResource is idle.

= Draw pass issue =
Compose does most of its measure/layout passes during a draw pass for
performance reasons. Since Robolectric doesn't perform draw passes, we
don't get measuring and layouting in Compose either.
To fix this issue, the RobolectricIdlingStrategy explicitly calls
`requestLayout()` after each call to Espresso.onIdle(), on each Compose
host that is still waiting for measure/layout. Robolectric still
performs measure/layout traversal on Views, so this will trigger
measure/layout in Compose as well.

Fix: 186203342
Test: ./gradlew compose:ui:ui-test:test
Relnote: "Added IdlingStrategy to AndroidComposeTestRule that can be
used by testing frameworks to install an alternative mechanism to await
or achieve quiescence. Use
`AndroidComposeTestRule.setIdlingStrategyFactory()` before your test
starts to install your own strategy."

Change-Id: I608fa541ffd0bfff6b847e873843df0701425529