Correctly handle interrution filter.
The interruption filter has a number of different nodes that,
depending on the device implementation, may correspond to the "Do Not
Disturb" state being set.

This commit updates the logic to support all those different modes and
defaults this value to the unknown value (which is part of the defined
API) until we get a value of this from the system.

Bug: 186503832
Test: ./gradlew :wear:wear-watchface:testDebugUnitTest
Change-Id: Iae344af831b2e5b7b6932a2bb8e77415881f6105