Wires target lang configuration to test methods
Changes `runProcessorTest` and similar methods to receive a
targetLanguage parameter which is set to JAVA by default. This is passed
down to the test backend through `TestCompilationParameters` class which
was modified to receive this configuration.

These changes reflected on the construction of `SyntheticKspProcessor`
used in tests which is no longer hardcoded to work with JAVA.