Add an API to get the fragment that was most recently added to a FragmentContainerView
Adding an API to get the fragment that was most recently added
to the calling FragmentContainerView. This allows users to get the
fragment directly from the FragmentContainerView, instead of having
to go through the FragmentManager.

RelNote: "`FragmentContainerView` now provides `getFragment()`,
which returns the fragment that was most recently added to the container."

Bug: 162527857
Test: FragmentContainerViewTest.kt
Change-Id: Ife17ac3a7afe345a8a2e8e1bdf9281fabd63a2d1