Don't return invalid schema from SchemaBundle
SchemaBundle returns whatever GSON parses from input without checking
whether it has a database or not. For some reason that i cannot
discover, this code starts not throwing when invoked with kotlin 1.5 in
classpath (or one of its dependencies).

Nevertheless, it seems to be the right thing to throw if the read schema
bundle does not include a database. I've changed it to throw if schema
file does not have a database and also added logging to the path where
we overwrite the schema.

We should do a followup to add some validation here but I'm only
fixing SchemaBundle for now to unblock kotlin 1.5 update.

I've also updated database processor to catch all exceptions when
parsing schema, not just encoding exception.

Bug: 187208773
Test: DatabaseProcessorTest

Change-Id: I32c7dc31600b5a0a06653b7a713e6af8b7915cf6