Enable emojicompat for AppCompatButton
This maintains the same API as AppCompatTextView using
isEmojiCompatEnabled/setEmojiCompatEnabled and a matching attribute.

This also introduces a common interface to all the EmojiCompat*Views in
AppCompat similar to the pattern used in AutoSizableTextView. This will
help ensure that the multiple EC configuration options are consistent
across views.

Relnote: "Add emojicompat integration to AppCompatButton similar
to AppCompatTextView. Added attribute `emojiCompatEnabled` as well as
getter and setter to configure the integration."

Test: New unit tests, integration with sample app
Bug: b/178035684
Change-Id: Ic262df79e5f11a5e3acb3a2078956834c84d8324