Fix flaky fragment animator test
The saveWhileAnimatingAway test attempts to do some checks while the
animator is running. This is not ALWAYS possible since some test
harnesses disable Animators we can end up not being able to do our
checks in time.

This changes the test to only do certain checks if the animator is
actually still running.

Test: RepeatTest 100 x
Bug: 187599370
Change-Id: I25e6da676ee0d778f3cfb0411994f5e5e86a0285