Add temporary workaround to pop back stack on sheet dismiss
NavigatorState's pop currently can't be called outside of popBackStackEntry; we need to pop the back stack outside of that though, e.g. when the sheet is no longer visible. BottomSheetNavigator now takes a onSheetDismissed lambda param that can encapsulate that logic so that the navigator doesn't have to depend on the NavController.

For convenience, rememberBottomSheetNavigator takes a NavController and calls NavController#popBackStack in the onSheetDismissed callback.

This will be removed ASAP; probably Navigation 2.4.0-alpha02. We'll provide a ReplaceWith quickfix then too.

See b/187873799 for a few more details on the issue.

Change-Id: Ia5da50279ef66dedea5b111638f336771574859a