Remove updateDefaultRequestBuilderToCameraControl
updateDefaultRequestBuilderToCameraControl is designed to be the
base 3A regions when comparing the 3A regions in CaptureResult
to check if startFocusMetering is finished or not.

Remove updateDefaultRequestBuilderToCameraControl by using
alternative approach: Adds auto-incremeted sessionUpdateId tag to
the SessionConfig so that it can check the id in CaptureResult
to know if updateSessionConfig is finished or not.

Bug: 148036411
Test: FocusMeteringControlTest, CameraControlDeviceTest
Change-Id: I4a979a64ba1c1df27478e545639f26e797e82ade