Fix tests for nested scrolling that bump into overscroll
Fixes: 187711104

When a RecyclerView hits the edge of the scrollable content
and the nested scroll doesn't accept the scroll event, the
overscroll now stretches. This means that further motion
events will continue to stretch instead of doing nested
scrolling. This plays havoc with tests that don't expect
that behavior, so I've disabled overscroll for these
tests.

Test: ran broken tests
Change-Id: I54619ff30bd5cf6b4a81f4b5f8541a8317e2ba5c