Changes Javac TestRunner target interaction
Explicitly states that JavacCompilationTestRunner is not supposed to
handle KOTLIN as a target language (and Kapt should be used instead).
The API surrounding it has been modified to not take that as a
parameter. If however it's still called to run on top of Kotlin code, it
will throw an error.