Setup room-paging artifact
room-paging will help us migrate dependencies on paging to its own
artifact so that users of room don't end up pulling paging in
unnecessarily. We'll also be using this artifact to host the native
PagingSource implementation for native paging3 support, which should
help alleviate a lot of the compat behavior errors and unnecesary
wrapping we currently do.

Relnote: "Added a new room-paging module"
Test: ./gradlew bOS
Change-Id: I592443bda12b54a036a231263544cd1a1f571476