Fix issues with stale surfaces and drawBlack
WallPaperService initially has a placeholder surfaceHolder which
later gets replaced by a real one. This is important for GL
rendering (which will fail with the placeholder). This patch makes
sure we wait for onSurfaceChanged before calling createWatchFace.

In addition we turn drawBlack into a NOP because we don't know
ahead of time whether the Renderer will use hardware or software
canvas. Skia crashes if we mix these so for now we turn this into
a NOP.

Test: Manual testing
Bug: 189452267
Change-Id: I1bf59462c305a9837691e20c6662ee602221f282