Do pre-measure as part of LazyColumn/Row prefetching
When prefetching was originally introduced it was mostly improving the cases with simple items, the ones which doesn't contain SubcomposeLayout(nested LazyRows or BoxWithConstraints). It happened because we were only precomposing the item in advance, not doing premeasuring for it. And SubcomposeLayout is postponing the real content composition till the measure stage.
With this change we will do premeasuring as well as part of prefetching. This improved NestedListsScrollBenchmark#frameTime99thPercentileMs by 28% on Pixel 3A.

Fixes: 187392865
Test: manually, LazyListPrefetcherTest, NestedListsScrollBenchmark
Change-Id: Ic3d1aa146b204185e2a3d49f2c6ec11952291642