Make LayoutNode insert/attach check folded parent
This changes the IllegalStateException checks in LayoutNode to be based
on the folded parent, for consistency with the logic which is based on
folded childen.

This is for diagnostic purposes: it cannot fix an underlying bug, but it
makes it more likely that an exception is thrown earlier, with a stack
trace indicating the root cause.  I also added debugTreeToString() for
additional diagnostics.

Context: in b/189892138 there is a bug where a parent has non-null owner
and its child has null owner, with a root cause I haven't figured out
yet.  These checks are supposed to guarantee this can't have happened,
and one potential gap in this guarantee is with unfolded children.

Bug: 189892138
Test: Existing tests
Change-Id: I068b130262d815082054bdc9b3bc9c1f76db4a76