Initial version of Card and AppCard.
Here is the initial version of the generic Wear Material Card component and with the
first of a small number of opinionated Cards called AppCard. It's expected there will
be a handful (2 or 3) other opinionated Cards which will follow that will also be
multi-slot with specific coloring of the slot contents. AppCard is like a notification
with specific areas of the card emphasised. AppCards body() slot could take either a
text area or an image.

Further golden image tests will follow once we have settled on the API.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 188887973
Change-Id: I5474a8f056ac635372d0d85a7f1d56740b7e2808