Run zip test apks tasks in server builds
buildOnServer does not build test apks anymore for library modules.

This PR adds the zip apks task to build them.

This is not fully ideal as it will zip apks into the dist (which already
gets zipped as github artifact). But using the same task has the added
benefit of staying close to the mainline.

Bug: n/a
Test: presubmit run outputs for the PR (on push target)