Bugfix: Don't reset inputType in AppCompatEditText
Calling setKeyListener in the AppCompatEditText constructor reset the
input type specified in XML. Instead, only on the constructor call
restore the input type after setting the EmojiCompat key listener.

The platfrom behavior is retained for AppCompatEditText.setKeyListener,
which will cause the inputType to be reset in setKeyListener on O+.

Bug: b/191061070
Test: New unit test
Relnote: "Fix bug in AppCompatEditText that will reset the inputType
specified in XML to remove variations. This bug was introduced in
AppCompat 1.4.0-alpha01."

Change-Id: I9df3655f05312806c46875e9d5603fe48154dd8e