Fix reentrant navigate/popBackStack from Lifeycle observers
When a developer adds a LifecycleObserver to a
NavBackStackEntry, the dispatch of those Lifecycle
change events is done by updateBackStackLifecycle().

Normally, this wouldn't be a problem, but if the
developer calls navigate() or popBackStack(), the
back stack will be changed during the dispatch,
leading to reentrant calls to both
dispatchOnDestinationChanged() and
updateBackStackLifecycle(), which can lead to improper
dispatch of changes and errors during
iteration of the back stack.

By hardening dispatchOnDestinationChanged() to
reentrant behavior and making
updateBackStackLifecycle() operate on a copy
of the list, we can avoid issues that occur due
to this reentrant behavior.

Relnote: "Fixed a `NoSuchElementException` when
calling `popBackStack()` from within a `LifecycleObserver`
attached to a `NavBackStackEntry` caused by
reentrant updates to the NavController's state."
Test: added tests to NavControllerTest suite
BUG: 192296981

Change-Id: I646215354c80b1c4571c7351b782db9897d1ddc2